requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib
python-multipart
torch
transformers
peft
sentencepiece
protobuf
huggingface_hub[hf_xet]
